Concatenate item description's first line onto base description
